opt: Update metadata files
Rename metadata_column.go to column_meta.go and create table_meta.go.
Move metadata functions to corresponding files.

This is in preparation for further refactoring of the Metadata types.

Release note: None
