+++ pkg/sql/opt/metadata.go
@@ -17,83 +17,11 @@ package opt
 import (
 	"context"
 	"fmt"
-	"strings"
 
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/cat"
 	"github.com/cockroachdb/cockroach/pkg/sql/privilege"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/types"
 )
-
-// TableID uniquely identifies the usage of a table within the scope of a
-// query. TableID 0 is reserved to mean "unknown table".
-//
-// Internally, the TableID consists of an index into the Metadata.tables slice,
-// as well as the ColumnID of the first column in the table. Subsequent columns
-// have sequential ids, relative to their ordinal position in the table.
-//
-// See the comment for Metadata for more details on identifiers.
-type TableID uint64
-
-const (
-	tableIDMask = 0xffffffff
-)
-
-// ColumnID returns the metadata id of the column at the given ordinal position
-// in the table.
-//
-// NOTE: This method does not do bounds checking, so it's up to the caller to
-//       ensure that a column really does exist at this ordinal position.
-func (t TableID) ColumnID(ord int) ColumnID {
-	return t.firstColID() + ColumnID(ord)
-}
-
-// makeTableID constructs a new TableID from its component parts.
-func makeTableID(index int, firstColID ColumnID) TableID {
-	// Bias the table index by 1.
-	return TableID((uint64(index+1) << 32) | uint64(firstColID))
-}
-
-// firstColID returns the ColumnID of the first column in the table.
-func (t TableID) firstColID() ColumnID {
-	return ColumnID(t & tableIDMask)
-}
-
-// index returns the index of the table in Metadata.tables. It's biased by 1, so
-// that TableID 0 can be be reserved to mean "unknown table".
-func (t TableID) index() int {
-	return int((t>>32)&tableIDMask) - 1
-}
-
-// TableAnnID uniquely identifies an annotation on an instance of table
-// metadata. A table annotation allows arbitrary values to be cached with table
-// metadata, which can be used to avoid recalculating base table properties or
-// other information each time it's needed.
-//
-// To create a TableAnnID, call NewTableAnnID during Go's program initialization
-// phase. The returned TableAnnID never clashes with other annotations on the
-// same table. Here is a usage example:
-//
-//   var myAnnID = NewTableAnnID()
-//
-//   md.SetTableAnnotation(TableID(1), myAnnID, "foo")
-//   ann := md.TableAnnotation(TableID(1), myAnnID)
-//
-// Currently, the following annotations are in use:
-//   - WeakKeys: weak keys derived from the base table
-//   - Stats: statistics derived from the base table
-//
-// To add an additional annotation, increase the value of maxTableAnnIDCount and
-// add a call to NewTableAnnID.
-type TableAnnID int
-
-// tableAnnIDCount counts the number of times NewTableAnnID is called.
-var tableAnnIDCount TableAnnID
-
-// maxTableAnnIDCount is the maximum number of times that NewTableAnnID can be
-// called. Calling more than this number of times results in a panic. Having
-// a maximum enables a static annotation array to be inlined into the metadata
-// table struct.
-const maxTableAnnIDCount = 2
 
 // Metadata assigns unique ids to the columns, tables, and other metadata used
 // within the scope of a particular query. Because it is specific to one query,
@@ -138,33 +66,6 @@ type mdDependency struct {
 	ds cat.DataSource
 
 	priv privilege.Kind
-}
-
-// mdTable stores information about one of the tables stored in the metadata.
-type mdTable struct {
-	// tab is a reference to the table in the catalog.
-	tab cat.Table
-
-	// anns annotates the table metadata with arbitrary data.
-	anns [maxTableAnnIDCount]interface{}
-}
-
-// mdColumn stores information about one of the columns stored in the metadata,
-// including its label and type.
-type mdColumn struct {
-	// tabID is the identifier of the base table to which this column belongs.
-	// If the column was synthesized (i.e. no base table), then the value is set
-	// to UnknownTableID.
-	tabID TableID
-
-	// label is the best-effort name of this column. Since the same column can
-	// have multiple labels (using aliasing), one of those is chosen to be used
-	// for pretty-printing and debugging. This might be different than what is
-	// stored in the physical properties and is presented to end users.
-	label string
-
-	// typ is the scalar SQL type of this column.
-	typ types.T
 }
 
 // Init prepares the metadata for use (or reuse).
@@ -242,111 +143,6 @@ func (md *Metadata) NumColumns() int {
 	return len(md.cols)
 }
 
-// IndexColumns returns the set of columns in the given index.
-// TODO(justin): cache this value in the table metadata.
-func (md *Metadata) IndexColumns(tableID TableID, indexOrdinal int) ColSet {
-	tab := md.Table(tableID)
-	index := tab.Index(indexOrdinal)
-
-	var indexCols ColSet
-	for i, cnt := 0, index.ColumnCount(); i < cnt; i++ {
-		ord := index.Column(i).Ordinal
-		indexCols.Add(int(tableID.ColumnID(ord)))
-	}
-	return indexCols
-}
-
-// ColumnTableID returns the identifier of the base table to which the given
-// column belongs. If the column has no base table because it was synthesized,
-// ColumnTableID returns zero.
-func (md *Metadata) ColumnTableID(id ColumnID) TableID {
-	// ColumnID is biased so that 0 is never used (reserved to indicate the
-	// unknown column).
-	return md.cols[id-1].tabID
-}
-
-// ColumnLabel returns the label of the given column. It is used for pretty-
-// printing and debugging.
-func (md *Metadata) ColumnLabel(id ColumnID) string {
-	// ColumnID is biased so that 0 is never used (reserved to indicate the
-	// unknown column).
-	return md.cols[id-1].label
-}
-
-// ColumnType returns the SQL scalar type of the given column.
-func (md *Metadata) ColumnType(id ColumnID) types.T {
-	// ColumnID is biased so that 0 is never used (reserved to indicate the
-	// unknown column).
-	return md.cols[id-1].typ
-}
-
-// ColumnOrdinal returns the ordinal position of the column in its base table.
-// It panics if the column has no base table because it was synthesized.
-func (md *Metadata) ColumnOrdinal(id ColumnID) int {
-	tabID := md.cols[id-1].tabID
-	if tabID == 0 {
-		panic("column was synthesized and has no ordinal position")
-	}
-	return int(id - tabID.firstColID())
-}
-
-// QualifiedColumnLabel returns the column label, qualified with the table name
-// if either of these conditions is true:
-//
-//   1. fullyQualify is true
-//   2. there's another column in the metadata with the same column name but
-//      different table name
-//
-// If the column label is qualified, the table is prefixed to it and separated
-// by a "." character. The table name is qualified with catalog/schema only if
-// fullyQualify is true.
-func (md *Metadata) QualifiedColumnLabel(id ColumnID, fullyQualify bool) string {
-	col := &md.cols[id-1]
-	if col.tabID == 0 {
-		// Column doesn't belong to a table, so no need to qualify it further.
-		return col.label
-	}
-	tab := md.Table(col.tabID)
-
-	// If a fully qualified label has not been requested, then only qualify it if
-	// it would otherwise be ambiguous.
-	var tabName string
-	if !fullyQualify {
-		for i := range md.cols {
-			if i == int(id-1) {
-				continue
-			}
-
-			// If there are two columns with same name, then column is ambiguous.
-			otherCol := &md.cols[i]
-			if otherCol.label == col.label {
-				tabName = string(tab.Name().TableName)
-				if otherCol.tabID == 0 {
-					fullyQualify = true
-				} else {
-					// Only qualify if the qualified names are actually different.
-					otherTabName := string(md.Table(otherCol.tabID).Name().TableName)
-					if tabName != otherTabName {
-						fullyQualify = true
-					}
-				}
-			}
-		}
-	} else {
-		tabName = tab.Name().FQString()
-	}
-
-	if !fullyQualify {
-		return col.label
-	}
-
-	var sb strings.Builder
-	sb.WriteString(tabName)
-	sb.WriteRune('.')
-	sb.WriteString(col.label)
-	return sb.String()
-}
-
 // AddTable indexes a new reference to a table within the query. Separate
 // references to the same table are assigned different table ids (e.g. in a
 // self-join query). All columns are added to the metadata. If mutation columns
@@ -390,40 +186,4 @@ func (md *Metadata) TableByStableID(id cat.StableID) cat.Table {
 		}
 	}
 	return nil
-}
-
-// TableAnnotation returns the given annotation that is associated with the
-// given table. If the table has no such annotation, TableAnnotation returns
-// nil.
-func (md *Metadata) TableAnnotation(tabID TableID, annID TableAnnID) interface{} {
-	return md.tables[tabID.index()].anns[annID]
-}
-
-// SetTableAnnotation associates the given annotation with the given table. The
-// annotation is associated by the given ID, which was allocated by
-// calling NewTableAnnID. If an annotation with the ID already exists on the
-// table, then it is overwritten.
-//
-// See the TableAnnID comment for more details and a usage example.
-func (md *Metadata) SetTableAnnotation(tabID TableID, tabAnnID TableAnnID, ann interface{}) {
-	md.tables[tabID.index()].anns[tabAnnID] = ann
-}
-
-// NewTableAnnID allocates a unique annotation identifier that is used to
-// associate arbitrary data with table metadata. Only maxTableAnnIDCount total
-// annotation ID's can exist in the system. Attempting to exceed the maximum
-// results in a panic.
-//
-// This method is not thread-safe, and therefore should only be called during
-// Go's program initialization phase (which uses a single goroutine to init
-// variables).
-//
-// See the TableAnnID comment for more details and a usage example.
-func NewTableAnnID() TableAnnID {
-	if tableAnnIDCount == maxTableAnnIDCount {
-		panic("can't allocate table annotation id; increase maxTableAnnIDCount to allow")
-	}
-	cnt := tableAnnIDCount
-	tableAnnIDCount++
-	return cnt
 }
